+++ src/pages/dialog/index.tsx
@@ -1,5 +1,5 @@
 import { View } from '@tarojs/components';
-import { useState } from 'react';
+import { useState, useRef } from 'react';
 import Dialog from '../../components/dialog';
 import Button from '../../components/button';
 import Typography from '../../components/typography';
@@ -10,6 +10,8 @@ const Index = () => {
   const [visible2, setVisible2] = useState(false);
   const [visible3, setVisible3] = useState(false);
   const [visible4, setVisible4] = useState(false);
+  const [visible5, setVisible5] = useState(false);
+  const dialogRef = useRef<any>(null);
   return (
     <>
       <View className='p-20'>
@@ -24,6 +26,36 @@ const Index = () => {
             message='逆水行舟，不进则退'
             onCancel={() => setVisible1(false)}
             onConfirm={() => setVisible1(false)}
+          />
+        </View>
+      </View>
+
+      <View className='p-20'>
+        <View>
+          <Button type='primary' onClick={() => setVisible5(true)}>
+            异步弹窗
+          </Button>
+          <Dialog
+            visible={visible5}
+            title='标题'
+            showCancelButton
+            message='逆水行舟，不进则退'
+            onCancel={() =>
+              new Promise((resolve) => {
+                setTimeout(() => {
+                  resolve(true);
+                  setVisible5(false);
+                }, 1200);
+              })
+            }
+            onConfirm={() =>
+              new Promise((resolve) => {
+                setTimeout(() => {
+                  resolve(true);
+                  setVisible5(false);
+                }, 2000);
+              })
+            }
           />
         </View>
       </View>
@@ -101,6 +133,28 @@ const Index = () => {
           >
             <Button type='primary'>Dialog.Alert</Button>
           </Dialog.Alert>
+        </View>
+      </View>
+
+      <View className='p-20'>
+        <Button type='primary' onClick={() => dialogRef.current.show()}>
+          DialogInstance
+        </Button>
+        <View>
+          <Dialog.Alert
+            title='标题'
+            ref={dialogRef}
+            theme='radio'
+            content={
+              <View style={{ margin: '14px' }}>
+                <Typography.Text>逆水行舟，不进则退。</Typography.Text>
+              </View>
+            }
+            showCancelButton
+            closeOnMaskClick
+            onCancel={() => console.log('onCancel')}
+            onConfirm={() => console.log('onConfirm')}
+          />
         </View>
       </View>
     </>